feat: Implement map functionality to visualize obras with markers and popups
- Added map.js to handle fetching obras and displaying them on a Leaflet map.
- Integrated geocoding to convert addresses into coordinates for marker placement.
- Created mapa.html template to render the map and display obra details.
- Developed veiculos_routes.py to manage vehicle-related routes, including listing, adding, editing, and deleting vehicles.
