+++ frontend/templates/default.html
@@ -1,63 +1,79 @@
 <!DOCTYPE html>
 <html lang="pt-br">
+
 <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>UaiRoute</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.6/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-4Q6Gf2aSP4eDXB8Miphtr37CMZZQ5oXLH2yaXMJ2w8e2ZtHTl7GptT4jmndRuHDT" crossorigin="anonymous">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"/>
-    {% block styles %}
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>UaiRoute</title>
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.6/dist/css/bootstrap.min.css" rel="stylesheet"
+    integrity="sha384-4Q6Gf2aSP4eDXB8Miphtr37CMZZQ5oXLH2yaXMJ2w8e2ZtHTl7GptT4jmndRuHDT" crossorigin="anonymous">
+  <link rel="stylesheet"
+    href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0" />
+  {% block styles %}
 
-    {% endblock %}
+  {% endblock %}
 </head>
+
 <body>
-    <nav class="navbar navbar-expand-lg bg-body-tertiary">
-        <div class="container-fluid">
-          <a class="navbar-brand" href="/">UaiRoute</a>
-          <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-            <span class="navbar-toggler-icon"></span>
-          </button>
-          <div class="collapse navbar-collapse" id="navbarNav">
-            <ul class="navbar-nav">
-                <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-                      Obras
-                    </a>
-                    <ul class="dropdown-menu">
-                      <li><a class="dropdown-item" href="/obras">Visualizar obras</a></li>
-                      <li><a class="dropdown-item" href="/obras/cadastrar">Cadastrar obra</a></li>
-                    </ul>
-                  </li>
-                  <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-                      Funcionários
-                    </a>
-                    <ul class="dropdown-menu">
-                      <li><a class="dropdown-item" href="/funcionarios">Visualizar funcionários</a></li>
-                      <li><a class="dropdown-item" href="/funcionarios/cadastrar">Cadastrar funcionário</a></li>
-                    </ul>
-                  </li>
-                  <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-                      Veículos
-                    </a>
-                    <ul class="dropdown-menu">
-                      <li><a class="dropdown-item" href="/veiculos">Visualizar veículos</a></li>
-                      <li><a class="dropdown-item" href="/veiculos/cadastrar">Cadastrar veículo</a></li>
-                    </ul>
-                  </li>
+  <nav class="navbar navbar-expand-lg bg-body-tertiary">
+    <div class="container-fluid">
+      <a class="navbar-brand" href="/">
+        <img src="{{ url_for('static', filename='images/logo.png') }}" alt="UaiRoute Logo" height="50" style="transform: scale(1.8);">
+      </a>
+      <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav"
+        aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+      </button>
+      <div class="collapse navbar-collapse" id="navbarNav">
+        <ul class="navbar-nav">
+          <li class="nav-item dropdown">
+            <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+              Obras
+            </a>
+            <ul class="dropdown-menu">
+              <li><a class="dropdown-item" href="/obras">Visualizar obras</a></li>
+              <li><a class="dropdown-item" href="/obras/cadastrar">Cadastrar obra</a></li>
             </ul>
-          </div>
-        </div>
-    </nav>
+          </li>
+          <li class="nav-item dropdown">
+            <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+              Funcionários
+            </a>
+            <ul class="dropdown-menu">
+              <li><a class="dropdown-item" href="/funcionarios">Visualizar funcionários</a></li>
+              <li><a class="dropdown-item" href="/funcionarios/cadastrar">Cadastrar funcionário</a></li>
+            </ul>
+          </li>
+          <li class="nav-item dropdown">
+            <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+              Veículos
+            </a>
+            <ul class="dropdown-menu">
+              <li><a class="dropdown-item" href="/veiculos">Visualizar veículos</a></li>
+              <li><a class="dropdown-item" href="/veiculos/cadastrar">Cadastrar veículo</a></li>
+            </ul>
+          </li>
+            <li class="nav-item d-flex align-items-center">
+              <a class="nav-link d-flex align-items-center gap-1 text-primary" href="/mapa" role="button" aria-expanded="false">
+                <span class="material-symbols-outlined" style="font-size: 22px; color: #0d6efd;">map</span>
+                Acessar o mapa
+              </a>
+            </li>
+        </ul>
+      </div>
+    </div>
+  </nav>
 
-    {% block conteudo %}
+  {% block conteudo %}
 
-    {% endblock %}
+  {% endblock %}
 
-    {% block scripts %}
+  {% block scripts %}
 
-    {% endblock %}
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.6/dist/js/bootstrap.bundle.min.js" integrity="sha384-j1CDi7MgGQ12Z7Qab0qlWQ/Qqz24Gc6BM0thvEMVjHnfYGF0rmFCozFSxQBxwHKO" crossorigin="anonymous"></script>
+  {% endblock %}
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.6/dist/js/bootstrap.bundle.min.js"
+    integrity="sha384-j1CDi7MgGQ12Z7Qab0qlWQ/Qqz24Gc6BM0thvEMVjHnfYGF0rmFCozFSxQBxwHKO"
+    crossorigin="anonymous"></script>
 </body>
+
 </html>
