+++ frontend/templates/home.html
@@ -82,7 +82,7 @@
                 <p class="mb-4" style="max-width: 520px; font-size:1.1rem;">
                     O UaiRoute é o seu copiloto para gestão de obras e logística: visualize, cadastre e otimize rotas entre obras, funcionários e veículos em poucos cliques. Reduza custos, ganhe tempo e tenha controle total sobre sua operação.
                 </p>
-                <a href="{{ url_for('obras') }}" class="btn btn-warning cta-btn shadow mb-5"><span class="material-symbols-outlined align-middle me-1">map</span> Explorar Obras</a>
+                <a href="{{ url_for('map') }}" class="btn btn-warning cta-btn shadow mb-5"><span class="material-symbols-outlined align-middle me-1">map</span> Explorar Mapa</a>
             </div>
         </div>
     </div>
